+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.25] - 2026-04-13
+
+### Changed 
+- Moved CoolProp as a provider, removed it as a core dependency. 
+- This means that it will be used as the default provider for units that require thermodynamic properties. 
+
 ## [0.2.24] - 2026-04-12
 
 ### Changed